Conflicts:
	app/src/main/java/com/worldventures/dreamtrips/modules/profile/presenter/ProfilePresenter.java
	app/src/main/java/com/worldventures/dreamtrips/modules/profile/view/fragment/ProfileFragment.java
	app/src/main/java/com/worldventures/dreamtrips/modules/trips/presenter/TripListPresenter.java
	app/src/main/res/layout/header_profile.xml
	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/worldventures/dreamtrips/modules/common/model/User.java
@@ -34,8 +34,8 @@ public class User extends BaseEntity implements Parcelable {
     private Date birthDate;
     private Date enrollDate;
 
-    private int dreamTripsPoints;
-    private int roviaBucks;
+    private double dreamTripsPoints;
+    private double roviaBucks;
     private int tripImagesCount;
     private int bucketListItemsCount;
 
@@ -110,11 +110,11 @@ public class User extends BaseEntity implements Parcelable {
         return enrollDate;
     }
 
-    public int getDreamTripsPoints() {
+    public double getDreamTripsPoints() {
         return dreamTripsPoints;
     }
 
-    public int getRoviaBucks() {
+    public double getRoviaBucks() {
         return roviaBucks;
     }
